refactor(mobile): fold the two export request registries into one
`HtmlExportRequests` was a near-verbatim copy of `MarkdownExportRequests`; the
only real difference is which key the success result carries, so that reader is
now a constructor argument and the pending map, the timeout and the docId check
are written once. Each call site still gets its own instance and its own result
type, which is what keeps one export from settling the other's promise.

The ported test covers both concrete usages, that a reply of the wrong kind
does not settle, and that HTML really does get the longer budget.
